# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  totalTrips: number;
  upcomingTrips: number;
  pastTrips: number;
  totalBudget: number;
  topDestinations: string[];
}

export interface UserProfile {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  language?: string;
  savedDestinations?: string[];
}

export interface UserProfileInput {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
  language?: string;
  savedDestinations?: string[];
}

export interface Trip {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  startDate: string;
  endDate: string;
  /** @nullable */
  coverPhotoUrl?: string | null;
  isPublic: boolean;
  /** @nullable */
  totalBudget?: number | null;
  createdAt: string;
  stopCount?: number;
}

export interface TripInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  startDate: string;
  endDate: string;
  coverPhotoUrl?: string;
  isPublic?: boolean;
  totalBudget?: number;
}

export interface TripUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  startDate?: string;
  endDate?: string;
  /** @nullable */
  coverPhotoUrl?: string | null;
  isPublic?: boolean;
  /** @nullable */
  totalBudget?: number | null;
}

export type TripBudgetBreakdown = {
  activities: number;
  accommodation: number;
  transport: number;
  meals: number;
};

export interface TripBudget {
  tripId: number;
  totalEstimated: number;
  /** @nullable */
  totalBudget: number | null;
  isOverBudget: boolean;
  perDayCost: number;
  breakdown: TripBudgetBreakdown;
}

export interface TripStop {
  id: number;
  tripId: number;
  cityId: number;
  cityName: string;
  cityCountry: string;
  /** @nullable */
  cityImageUrl?: string | null;
  startDate: string;
  endDate: string;
  order: number;
  /** @nullable */
  notes?: string | null;
  activityCount?: number;
}

export interface TripStopInput {
  cityId: number;
  startDate: string;
  endDate: string;
  order: number;
  notes?: string;
}

export interface TripStopUpdate {
  startDate?: string;
  endDate?: string;
  order?: number;
  /** @nullable */
  notes?: string | null;
}

export interface Activity {
  id: number;
  cityId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  type: string;
  estimatedCost: number;
  duration: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface StopActivity {
  id: number;
  stopId: number;
  activityId: number;
  activityName: string;
  activityType: string;
  /** @nullable */
  activityImageUrl?: string | null;
  /** @nullable */
  scheduledTime?: string | null;
  /** @nullable */
  notes?: string | null;
  cost: number;
  duration?: number;
}

export interface StopActivityInput {
  activityId: number;
  scheduledTime?: string;
  notes?: string;
  cost?: number;
}

export interface PackingItem {
  id: number;
  tripId: number;
  name: string;
  category: string;
  isPacked: boolean;
}

export interface PackingItemInput {
  /** @minLength 1 */
  name: string;
  category: string;
  isPacked?: boolean;
}

export interface PackingItemUpdate {
  name?: string;
  category?: string;
  isPacked?: boolean;
}

export interface TripNote {
  id: number;
  tripId: number;
  /** @nullable */
  stopId?: number | null;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface TripNoteInput {
  /** @minLength 1 */
  content: string;
  stopId?: number;
}

export interface TripNoteUpdate {
  /** @minLength 1 */
  content?: string;
  /** @nullable */
  stopId?: number | null;
}

export interface City {
  id: number;
  name: string;
  country: string;
  region: string;
  costIndex: number;
  popularity: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface PublicTrip {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverPhotoUrl?: string | null;
  startDate: string;
  endDate: string;
  authorName?: string;
  stops: TripStop[];
}