# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Traveloop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips/{tripId}/stops/{stopId}/activities:
    get:
      operationId: listStopActivities
      tags: [activities]
      summary: List activities for a stop
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of stop activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StopActivity"
    post:
      operationId: addStopActivity
      tags: [activities]
      summary: Add an activity to a stop
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopActivityInput"
      responses:
        "201":
          description: Added stop activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StopActivity"
  /trips/{tripId}/stops/{stopId}/activities/{stopActivityId}:
    delete:
      operationId: removeStopActivity
      tags: [activities]
      summary: Remove an activity from a stop
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
        - name: stopActivityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Packing ────────────────────────────────────────────────────────────────
  /trips/{tripId}/packing:
    get:
      operationId: listPackingItems
      tags: [packing]
      summary: List packing items for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of packing items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackingItem"
    post:
      operationId: createPackingItem
      tags: [packing]
      summary: Add a packing item to a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackingItemInput"
      responses:
        "201":
          description: Created packing item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingItem"
  /trips/{tripId}/packing/{itemId}:
    patch:
      operationId: updatePackingItem
      tags: [packing]
      summary: Update a packing item
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackingItemUpdate"
      responses:
        "200":
          description: Updated packing item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingItem"
    delete:
      operationId: deletePackingItem
      tags: [packing]
      summary: Delete a packing item
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Notes ──────────────────────────────────────────────────────────────────
  /trips/{tripId}/notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripNote"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a trip note
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripNoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripNote"
  /trips/{tripId}/notes/{noteId}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a trip note
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripNoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripNote"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a trip note
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Cities ─────────────────────────────────────────────────────────────────
...
  /cities/{cityId}:
    get:
      operationId: getCity
      tags: [cities]
      summary: Get a city by ID
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: City
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
  /cities/{cityId}/activities:
    get:
      operationId: listCityActivities
      tags: [cities]
      summary: List activities for a city
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

  # ── Public ─────────────────────────────────────────────────────────────────
  /public/trips/{tripId}:
    get:
      operationId: getPublicTrip
      tags: [public]
      summary: Get a public trip itinerary
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public trip view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTrip"
        "404":
          description: Not found or not public
components:
  schemas:
...
    TripStop:
      type: object
      required: [id, tripId, cityId, cityName, cityCountry, startDate, endDate, order]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        cityId:
          type: integer
        cityName:
          type: string
        cityCountry:
          type: string
        cityImageUrl:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        order:
          type: integer
        notes:
          type: ["string", "null"]
        activityCount:
          type: integer
...
    Activity:
      type: object
      required: [id, cityId, name, type, estimatedCost, duration]
      properties:
        id:
          type: integer
        cityId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
        estimatedCost:
          type: number
        duration:
          type: number
        imageUrl:
          type: ["string", "null"]
    StopActivity:
      type: object
      required: [id, stopId, activityId, activityName, activityType, cost]
      properties:
        id:
          type: integer
        stopId:
          type: integer
        activityId:
          type: integer
        activityName:
          type: string
        activityType:
          type: string
        activityImageUrl:
          type: ["string", "null"]
        scheduledTime:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        cost:
          type: number
        duration:
          type: number
    StopActivityInput:
      type: object
      required: [activityId]
      properties:
        activityId:
          type: integer
        scheduledTime:
          type: string
        notes:
          type: string
        cost:
          type: number

    # ── Packing ───────────────────────────────────────────────────────────────
    PackingItem:
      type: object
      required: [id, tripId, name, category, isPacked]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        name:
          type: string
        category:
          type: string
        isPacked:
          type: boolean
    PackingItemInput:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        isPacked:
          type: boolean
    PackingItemUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        isPacked:
          type: boolean

    # ── Notes ─────────────────────────────────────────────────────────────────
    TripNote:
      type: object
      required: [id, tripId, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        stopId:
          type: ["integer", "null"]
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    TripNoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        stopId:
          type: integer
    TripNoteUpdate:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        stopId:
          type: ["integer", "null"]

    # ── City ──────────────────────────────────────────────────────────────────
    City:
      type: object
      required: [id, name, country, region, costIndex, popularity]
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        region:
          type: string
        costIndex:
          type: number
        popularity:
          type: integer
        imageUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]

    # ── Public ────────────────────────────────────────────────────────────────
    PublicTrip:
      type: object
      required: [id, name, startDate, endDate, stops]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        coverPhotoUrl:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        authorName:
          type: string
        stops:
          type: array
          items:
            $ref: "#/components/schemas/TripStop"