# lib/api-zod/src/generated/types/stopActivity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StopActivity {
  id: number;
  stopId: number;
  activityId: number;
  activityName: string;
  activityType: string;
  /** @nullable */
  activityImageUrl?: string | null;
  /** @nullable */
  scheduledTime?: string | null;
  /** @nullable */
  notes?: string | null;
  cost: number;
  duration?: number;
}