# lib/api-zod/src/generated/types/stopActivityInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StopActivityInput {
  activityId: number;
  scheduledTime?: string;
  notes?: string;
  cost?: number;
}